+++ examples/demo-rollup/src/celestia_rollup.rs
@@ -1,11 +1,8 @@
-use std::str::FromStr;
-
 use async_trait::async_trait;
 use backon::ExponentialBuilder;
 use demo_stf::authentication::ModAuth;
 use demo_stf::genesis_config::StorageConfig;
 use demo_stf::runtime::Runtime;
-use sha2::Sha256;
 use sov_celestia_adapter::verifier::{CelestiaSpec, CelestiaVerifier, RollupParams};
 use sov_celestia_adapter::{CelestiaConfig, CelestiaService};
 use sov_db::ledger_db::LedgerDb;
@@ -13,7 +10,7 @@ use sov_kernels::basic::BasicKernel;
 use sov_mock_zkvm::{MockCodeCommitment, MockZkVerifier, MockZkvm};
 use sov_modules_api::default_spec::DefaultSpec;
 use sov_modules_api::execution_mode::{ExecutionMode, Native, Zk};
-use sov_modules_api::{Address, CryptoSpec, Spec};
+use sov_modules_api::{CryptoSpec, Spec};
 use sov_modules_rollup_blueprint::pluggable_traits::PluggableSpec;
 use sov_modules_rollup_blueprint::{FullNodeBlueprint, RollupBlueprint, WalletBlueprint};
 use sov_modules_stf_blueprint::{RuntimeEndpoints, StfBlueprint};
@@ -28,7 +25,7 @@ use sov_state::{DefaultStorageSpec, Storage, ZkStorage};
 use sov_stf_runner::{ParallelProverService, ProverService, RollupConfig, RollupProverConfig};
 use tokio::sync::watch;
 
-use crate::{PROVER_ADDRESS, ROLLUP_BATCH_NAMESPACE, ROLLUP_PROOF_NAMESPACE};
+use crate::{ROLLUP_BATCH_NAMESPACE, ROLLUP_PROOF_NAMESPACE};
 
 /// Rollup with CelestiaDa
 #[derive(Default)]
@@ -86,7 +83,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
         ledger_db: &LedgerDb,
         sequencer_db: &SequencerDb,
         da_service: &Self::DaService,
-        rollup_config: &RollupConfig<Self::DaConfig>,
+        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
     ) -> anyhow::Result<RuntimeEndpoints> {
         let sequencer = rollup_config.da.own_celestia_address.clone();
 
@@ -115,7 +112,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
 
     async fn create_da_service(
         &self,
-        rollup_config: &RollupConfig<Self::DaConfig>,
+        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
     ) -> Self::DaService {
         DaServiceWithRetries::with_exponential_backoff(
             CelestiaService::new(
@@ -135,7 +132,7 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
     async fn create_prover_service(
         &self,
         prover_config: RollupProverConfig,
-        _rollup_config: &RollupConfig<Self::DaConfig>,
+        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
         _da_service: &Self::DaService,
     ) -> Self::ProverService {
         let inner_vm = Risc0Host::new(risc0::ROLLUP_ELF);
@@ -157,13 +154,13 @@ impl FullNodeBlueprint<Native> for CelestiaDemoRollup<Native> {
             prover_config,
             zk_storage,
             CodeCommitment::default(),
-            Address::<Sha256>::from_str(PROVER_ADDRESS).expect("Prover address is not valid"),
+            rollup_config.proof_manager.prover_address,
         )
     }
 
     fn create_storage_manager(
         &self,
-        rollup_config: &RollupConfig<Self::DaConfig>,
+        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
     ) -> Result<Self::StorageManager, anyhow::Error> {
         let storage_config = StorageConfig {
             path: rollup_config.storage.path.clone(),